requirements.txt
# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
transformers>=4.35.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# NLP
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0
sentence-transformers>=2.2.0

# Time Series
statsmodels>=0.14.0

# Data Processing
pyyaml>=6.0
python-dateutil>=2.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Model Interpretability
shap>=0.43.0
lime>=0.2.0

# Experiment Tracking
mlflow>=2.8.0

# Web Scraping (if needed)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
